fix: allow prop-types to be removed in production
fixes #257
