+++ src/components/picker/picker.js
@@ -3,7 +3,8 @@ import React from 'react'
 import data from '../../../data/all.json'
 import NimblePicker from './nimble-picker'
 
-import { PickerPropTypes, PickerDefaultProps } from '../../utils/shared-props'
+import { PickerPropTypes } from '../../utils/shared-props'
+import { PickerDefaultProps } from '../../utils/shared-default-props'
 
 export default class Picker extends React.PureComponent {
   render() {
@@ -11,5 +12,5 @@ export default class Picker extends React.PureComponent {
   }
 }
 
-Picker.propTypes = PickerPropTypes
+Picker.propTypes /* remove-proptypes */ = PickerPropTypes
 Picker.defaultProps = { ...PickerDefaultProps, data }
